- Main Area bug fixed, crashing bugs resolved. - vieworder, vieworderpage, vieworderpagetablet from component/cart_widget.dart organize into Firebase/view_order.dart - Some of pending main area already apply (not function yet)
