+++ lib/Firebase/view_order.dart
@@ -1,10 +1,11 @@
+import 'package:firebase_database/firebase_database.dart';
+import 'package:flutter/material.dart';
 import 'package:fypmerchant/Color/color.dart';
 import 'package:fypmerchant/Components/button_widget.dart';
 import 'package:fypmerchant/Firebase/update_data.dart';
 import 'package:firebase_database/firebase_database.dart';
 import 'package:firebase_database/ui/firebase_animated_list.dart';
 import 'package:firebase_storage/firebase_storage.dart';
-import 'package:flutter/material.dart';
 
 class ViewOrder extends StatefulWidget {
   final double? maxWidth; // Declare maxWidth parameter
@@ -14,7 +15,6 @@ class ViewOrder extends StatefulWidget {
   @override
   State<ViewOrder> createState() => _ViewOrderState();
 }
-
 class _ViewOrderState extends State<ViewOrder> {
   final DatabaseReference query = FirebaseDatabase.instance.ref().child('Order');
 
@@ -96,9 +96,8 @@ class _ViewOrderState extends State<ViewOrder> {
       ],
     );
   }
-  // Function to show the AlertDialog
+// Function to show the AlertDialog
 }
-
 
 class ShowOrderPage extends StatefulWidget {
   final String username;
@@ -108,7 +107,6 @@ class ShowOrderPage extends StatefulWidget {
   @override
   State<ShowOrderPage> createState() => _ShowOrderPageState();
 }
-
 class _ShowOrderPageState extends State<ShowOrderPage> {
   late DatabaseReference query;
   String? username;
@@ -290,7 +288,6 @@ class ShowOrderPageTablet extends StatefulWidget {
   @override
   State<ShowOrderPageTablet> createState() => _ShowOrderPageTabletState();
 }
-
 class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
   late DatabaseReference query;
   String? username;
@@ -302,18 +299,17 @@ class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
   @override
   void initState() {
     super.initState();
-    String username = widget.username;
-    query = FirebaseDatabase.instance.ref().child('Order/${widget.username}');
+    username = widget.username; // Assigning widget's username to local username
+    query = FirebaseDatabase.instance.ref().child('Order/$username');
   }
 
   Future<String?> getImageUrl(String imageName) async {
     try {
-      Reference imageRef = FirebaseStorage.instance.ref().child(
-          'Item/$imageName');
+      Reference imageRef = FirebaseStorage.instance.ref().child('Item/$imageName');
       String imageUrl = await imageRef.getDownloadURL();
       return imageUrl;
     } catch (e) {
-      // print("Error retrieving image URL: $e");
+      // Handle error
       return null;
     }
   }
@@ -321,7 +317,7 @@ class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
   // Method to handle accepting all orders
   void acceptAllOrders() {
     for (final order in orders) {
-      AcceptOrder().readCartData(widget.username);
+      AcceptOrder().readCartData(username!);
     }
     setState(() {
       showOrder = false;
@@ -331,7 +327,7 @@ class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
   // Method to handle rejecting all orders
   void rejectAllOrders() {
     for (final order in orders) {
-      RejectOrder().rejectUserOrder(widget.username);
+      RejectOrder().rejectUserOrder(username);
     }
     setState(() {
       showOrder = false;
@@ -363,15 +359,13 @@ class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
                             return FutureBuilder<String?>(
                               future: getImageUrl(data['itemImage']),
                               builder: (context, snapshot) {
-                                if (snapshot.connectionState ==
-                                    ConnectionState.waiting) {
+                                if (snapshot.connectionState == ConnectionState.waiting) {
                                   // Show a placeholder while waiting for the image URL
                                   return const CircularProgressIndicator();
                                 }
                                 imageUrl = snapshot.data ??
                                     'https://fastly.picsum.photos/id/237/536/354.jpg?hmac=i0yVXW1ORpyCZpQ-CknuyV-jbtU7_x9EBQVhvT5aRr0';
-                                String itemName = data['itemName']?.toString() ??
-                                    '';
+                                String itemName = data['itemName']?.toString() ?? '';
                                 double price = data['price']?.toDouble() ?? 0.0;
                                 int qty = data['qty'] as int? ?? 0;
 
@@ -382,15 +376,12 @@ class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
                                       borderRadius: BorderRadius.circular(10.0),
                                     ),
                                     child: Row(
-                                      crossAxisAlignment: CrossAxisAlignment
-                                          .start,
+                                      crossAxisAlignment: CrossAxisAlignment.start,
                                       children: [
                                         Padding(
-                                          padding: const EdgeInsets.fromLTRB(
-                                              10, 10, 0, 10),
+                                          padding: const EdgeInsets.fromLTRB(10, 10, 0, 10),
                                           child: ClipRRect(
-                                            borderRadius: BorderRadius.circular(
-                                                5),
+                                            borderRadius: BorderRadius.circular(5),
                                             child: Image.network(
                                               imageUrl!,
                                               width: 130,
@@ -399,18 +390,14 @@ class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
                                             ),
                                           ),
                                         ),
-                                        const Padding(
-                                            padding: EdgeInsets.only(left: 10)),
+                                        const Padding(padding: EdgeInsets.only(left: 10)),
                                         SizedBox(
                                           width: 130,
                                           child: Padding(
-                                            padding: const EdgeInsets.fromLTRB(
-                                                10, 10, 0, 10),
+                                            padding: const EdgeInsets.fromLTRB(10, 10, 0, 10),
                                             child: Column(
-                                              mainAxisAlignment: MainAxisAlignment
-                                                  .start,
-                                              crossAxisAlignment: CrossAxisAlignment
-                                                  .start,
+                                              mainAxisAlignment: MainAxisAlignment.start,
+                                              crossAxisAlignment: CrossAxisAlignment.start,
                                               children: <Widget>[
                                                 Text(
                                                   itemName,
@@ -424,12 +411,10 @@ class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
                                                   style: const TextStyle(
                                                     fontSize: 16,
                                                     fontWeight: FontWeight.normal,
-                                                    color: CustomColors
-                                                        .primaryColor,
+                                                    color: CustomColors.primaryColor,
                                                   ),
                                                 ),
-                                                const Padding(
-                                                    padding: EdgeInsets.all(10)),
+                                                const Padding(padding: EdgeInsets.all(10)),
                                                 Text(
                                                   qty.toString(),
                                                   style: const TextStyle(
@@ -449,7 +434,6 @@ class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
                               },
                             );
                           } else {
-                            // Handle other cases, for example, if the data is not a Map
                             return Container();
                           }
                         },
@@ -468,7 +452,6 @@ class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
                     padding: EdgeInsets.all(20),
                     child: ButtonWidget.buttonWidget("Accept", () => acceptAllOrders()),
                   ),
-
                   Padding(
                     padding: EdgeInsets.all(20),
                     child: ButtonWidget.buttonWidget("Reject", () => rejectAllOrders()),
@@ -486,4 +469,3 @@ class _ShowOrderPageTabletState extends State<ShowOrderPageTablet> {
     );
   }
 }
-
